Split your data into a training and testing set

Create a training and testing data set. Also returns a bootstrapped version of the training data set.

trainTestSplit(
  data = df,
  splitAmt = 0.8,
  timeDependent = FALSE,
  responseVar = "nameOfResponseVar",
  stratifyOnResponse = FALSE,
  numberOfBootstrapSamples = 25
)

Arguments

data
The data set of interest.

splitAmt
The amount of data you want in the training set. Default is .8

timeDependent
Logical. Is your data time-dependent? If so, set TRUE.

responseVar
Name of response variable in analysis.

stratifyOnResponse
Logical. Should the training and testing splits be stratified based on the response? If so, set TRUE.

numberOfBootstrapSamples
Numeric. How many bootstrap samples do you want? Default is 25.

Value

A list with four components: train is the training set, test is the testing set, boot is a bootstrapped data set, and split is an rsample object that helps split your original data set.

Examples

library(easytidymodels)
library(dplyr)
utils::data(penguins, package = "modeldata")

#> Error in find.package(package, lib.loc, verbose = verbose): there is no package called ‘modeldata’

resp <- "sex"
split <- trainTestSplit(penguins, stratifyOnResponse = TRUE, responseVar = resp)

#> Error in eval_select_impl(NULL, .vars, expr(c(!!!dots)), include = .include,     exclude = .exclude, strict = .strict, name_spec = unique_name_spec,     uniquely_named = TRUE): object 'penguins' not found

#Training data
split$train

#> Error in split$train: object of type 'closure' is not subsettable

#Testing data
split$test

#> Error in split$test: object of type 'closure' is not subsettable

#Bootstrapped data
split$boot

#> Error in split$boot: object of type 'closure' is not subsettable

#Split object (helpful to call if you want to do model stacking)
split$split

#> Error in split$split: object of type 'closure' is not subsettable