Split your data into a training and testing set

Create a training and testing data set. Also returns a bootstrapped version of the training data set.

trainTestSplit(
  data = df,
  splitAmt = 0.8,
  timeDependent = FALSE,
  responseVar = "nameOfResponseVar",
  stratifyOnResponse = FALSE,
  numberOfBootstrapSamples = 25
)

Arguments

data
The data set of interest.

splitAmt
The amount of data you want in the training set. Default is .8

timeDependent
Logical. Is your data time-dependent? If so, set TRUE.

responseVar
Name of response variable in analysis.

stratifyOnResponse
Logical. Should the training and testing splits be stratified based on the response? If so, set TRUE.

numberOfBootstrapSamples
Numeric. How many bootstrap samples do you want? Default is 25.

Value

A list with four components: train is the training set, test is the testing set, boot is a bootstrapped data set, and split is an rsample object that helps split your original data set.

Examples

library(easytidymodels)
library(dplyr)
utils::data(penguins, package = "modeldata")
resp <- "sex"
split <- trainTestSplit(penguins, stratifyOnResponse = TRUE, responseVar = resp)
#Training data
split$train

#> # A tibble: 276 x 7
#>    species island bill_length_mm bill_depth_mm flipper_length_… body_mass_g
#>    <fct>   <fct>           <dbl>         <dbl>            <int>       <int>
#>  1 Adelie  Torge…           39.5          17.4              186        3800
#>  2 Adelie  Torge…           40.3          18                195        3250
#>  3 Adelie  Torge…           NA            NA                 NA          NA
#>  4 Adelie  Torge…           36.7          19.3              193        3450
#>  5 Adelie  Torge…           39.3          20.6              190        3650
#>  6 Adelie  Torge…           39.2          19.6              195        4675
#>  7 Adelie  Torge…           34.1          18.1              193        3475
#>  8 Adelie  Torge…           42            20.2              190        4250
#>  9 Adelie  Torge…           37.8          17.1              186        3300
#> 10 Adelie  Torge…           37.8          17.3              180        3700
#> # … with 266 more rows, and 1 more variable: sex <fct>

#Testing data
split$test

#> # A tibble: 68 x 7
#>    species island bill_length_mm bill_depth_mm flipper_length_… body_mass_g
#>    <fct>   <fct>           <dbl>         <dbl>            <int>       <int>
#>  1 Adelie  Torge…           39.1          18.7              181        3750
#>  2 Adelie  Torge…           38.9          17.8              181        3625
#>  3 Adelie  Torge…           46            21.5              194        4200
#>  4 Adelie  Biscoe           35.3          18.9              187        3800
#>  5 Adelie  Biscoe           40.6          18.6              183        3550
#>  6 Adelie  Biscoe           40.5          17.9              187        3200
#>  7 Adelie  Dream            39.5          16.7              178        3250
#>  8 Adelie  Dream            37.2          18.1              178        3900
#>  9 Adelie  Dream            36.4          17                195        3325
#> 10 Adelie  Dream            39.2          21.1              196        4150
#> # … with 58 more rows, and 1 more variable: sex <fct>

#Bootstrapped data
split$boot

#> # Bootstrap sampling 
#> # A tibble: 25 x 2
#>    splits            id         
#>    <list>            <chr>      
#>  1 <split [276/107]> Bootstrap01
#>  2 <split [276/104]> Bootstrap02
#>  3 <split [276/107]> Bootstrap03
#>  4 <split [276/94]>  Bootstrap04
#>  5 <split [276/102]> Bootstrap05
#>  6 <split [276/92]>  Bootstrap06
#>  7 <split [276/97]>  Bootstrap07
#>  8 <split [276/103]> Bootstrap08
#>  9 <split [276/100]> Bootstrap09
#> 10 <split [276/107]> Bootstrap10
#> # … with 15 more rows

#Split object (helpful to call if you want to do model stacking)
split$split

#> <Analysis/Assess/Total>
#> <276/68/344>